+++ core/src/main/java/org/elasticsearch/index/NodeServicesProvider.java
@@ -20,10 +20,8 @@
 package org.elasticsearch.index;
 
 import org.elasticsearch.client.Client;
-import org.elasticsearch.common.Nullable;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.util.BigArrays;
-import org.elasticsearch.indices.IndicesWarmer;
 import org.elasticsearch.indices.breaker.CircuitBreakerService;
 import org.elasticsearch.indices.cache.query.IndicesQueryCache;
 import org.elasticsearch.indices.query.IndicesQueriesRegistry;
@@ -39,7 +37,6 @@ public final class NodeServicesProvider {
 
     private final ThreadPool threadPool;
     private final IndicesQueryCache indicesQueryCache;
-    private final IndicesWarmer warmer;
     private final BigArrays bigArrays;
     private final Client client;
     private final IndicesQueriesRegistry indicesQueriesRegistry;
@@ -47,10 +44,9 @@ public final class NodeServicesProvider {
     private final CircuitBreakerService circuitBreakerService;
 
     @Inject
-    public NodeServicesProvider(ThreadPool threadPool, IndicesQueryCache indicesQueryCache, @Nullable IndicesWarmer warmer, BigArrays bigArrays, Client client, ScriptService scriptService, IndicesQueriesRegistry indicesQueriesRegistry, CircuitBreakerService circuitBreakerService) {
+    public NodeServicesProvider(ThreadPool threadPool, IndicesQueryCache indicesQueryCache, BigArrays bigArrays, Client client, ScriptService scriptService, IndicesQueriesRegistry indicesQueriesRegistry, CircuitBreakerService circuitBreakerService) {
         this.threadPool = threadPool;
         this.indicesQueryCache = indicesQueryCache;
-        this.warmer = warmer;
         this.bigArrays = bigArrays;
         this.client = client;
         this.indicesQueriesRegistry = indicesQueriesRegistry;
@@ -64,10 +60,6 @@ public final class NodeServicesProvider {
 
     public IndicesQueryCache getIndicesQueryCache() {
         return indicesQueryCache;
-    }
-
-    public IndicesWarmer getWarmer() {
-        return warmer;
     }
 
     public BigArrays getBigArrays() { return bigArrays; }
